Fix Hardsigmoid configuration in SE layer
In PyTorch there are no parameters to further tune the Hardsigmoid
activation function, in Burn one can set "alpha" and "beta" values
which are 0.2 and 0.5 by default.

While beta matches the PyTorch implementation, alpha does not, this
caused our values to drift away when inferencing through the layers,
resulting in a wrong confidence score.

The PyTorch alpha equivalent is 1/6 and configuring our Hardsigmoid
function with this value fixed the issue.